requirements.txt
langchain
langchain_core
langgraph
langchain-google-genai
yfinance
ddgs
python-dotenv
fastapi
uvicorn